Search and Expand Tags Based on Indices

This function provides a searchable table of tags and names of of V-Dem indices included in vdemlite. The table is based on the Structure of V-Dem Indices, Components and Indicators document and includes four columns: the variable tag; the index or indicator descriptor; the level of the index or indicator (high-level index, mid-level index, lower-level index; or indicator); and a column called "Part" that distinguishes between V-Dem Indicators and Indices (part 1) and indices created using V-Dem data (part 2). The default behavior is to return all indices and their descriptions. If a character vector of indices is provided, the function will return any sub-indices and indicators used to construct the selected indices.

Usage

searchdem(indicator = NULL)

Arguments

indicator

A character vector of indices to search and expand. Useful when looking for sub-indices and indicators used to construct a high-level index. Default is NULL.

Value

An interactive table with names and descriptions of V-Dem indicators and indices included in the vdemlite dataset.

Examples

# Return the full table of indices
searchdem()

  

    
Variable Tags and Descriptions

    

  

  

  

# Return indices and indicators used to construct polyarchy
searchdem("v2x_polyarchy")

  

    
Variable Tags and Descriptions

    

  

  

  

# Return indicators used to construct women's political empowerment and rule of law indices
searchdem(c("v2x_gender", "v2x_rule"))

  

    
Variable Tags and Descriptions